refactor: 같은 역할을 하는 Lotto, LottoNumbers 합치도록 변경
