+++ src/main/kotlin/lotto/model/LottoResult.kt
@@ -6,7 +6,7 @@ class LottoResult(
 ) {
 
     companion object {
-        fun of(lottoList: List<Lotto>, winLottoNumbers: LottoNumbers): LottoResult {
+        fun of(lottoList: List<Lotto>, winLottoNumbers: Lotto): LottoResult {
             val lottoRankList = lottoList.map { it.matches(winLottoNumbers) }.map { LottoRank.find(it) }
             val profitRate = calculateProfitRate(lottoRankList)
             return LottoResult(
