+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,7 +1,7 @@
 package lotto.view
 
+import lotto.model.Lotto
 import lotto.model.LottoNumber
-import lotto.model.LottoNumbers
 
 class InputView {
 
@@ -10,13 +10,13 @@ class InputView {
         return readln().toInt()
     }
 
-    fun inputLastWinningNumbers(): LottoNumbers {
+    fun inputLastWinningNumbers(): Lotto {
         println("지난 주 당첨 번호를 입력해 주세요.")
         return readln()
             .split(DEFAULT_DELIMITER)
             .map { it.trim() }
             .map { LottoNumber.from(it.toInt()) }
-            .let(::LottoNumbers)
+            .let(::Lotto)
     }
 
     companion object {
